+++ aeon/dj_pipeline/acquisition.py
@@ -11,7 +11,7 @@ from aeon.analysis import utils as analysis_utils
 
 from . import get_schema_name
 from .utils import paths
-from .ingest.load_metadata import extract_epoch_metadata, ingest_epoch_metadata
+from .populate.load_metadata import extract_epoch_metadata, ingest_epoch_metadata
 
 
 schema = dj.schema(get_schema_name("acquisition"))
@@ -1042,7 +1042,7 @@ def _load_legacy_subjectdata(experiment_name, data_dir, start, end):
         return subject_data
 
     if experiment_name == "social0-r1":
-        from aeon.dj_pipeline.ingest.create_socialexperiment_0 import fixID
+        from aeon.dj_pipeline.populate.create_socialexperiment_0 import fixID
 
         sessdf = subject_data.copy()
         sessdf = sessdf[~sessdf.id.str.contains("test")]
