+++ aeon/dj_pipeline/populate/create_experiment_02.py
@@ -1,7 +1,7 @@
 from aeon.dj_pipeline import acquisition, lab, subject
 
 
-# ============ Manual and automatic steps to for experiment 0.2 ingest ============
+# ============ Manual and automatic steps to for experiment 0.2 populate ============
 experiment_name = "exp0.2-r0"
 _weight_scale_rate = 20
